CI: fix using `PHP_CS_FIXER_FAST_LINT_TEST_CASES`
